♻️ refactor: openAiResponse 패키지 변경
