+++ app/src/main/kotlin/com/firstpenguin/app/domain/batch/dto/ai/OpenAiFileResponse.kt
@@ -1,4 +1,4 @@
-package com.firstpenguin.app.domain.batch.dto
+package com.firstpenguin.app.domain.batch.dto.ai
 
 data class OpenAiFileResponse(
     val id: String,
